pyproject.toml
[project]
name = "elastic-search-demo"
version = "0.1.0"
description = "A demo project for Elastic Search with semantic search capabilities"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "elasticsearch>=8.10.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.2",
    "httpx>=0.25.0",
    "numpy>=1.24.3",
    "pandas>=2.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["data", "backend", "frontend"] 